♻️ Move QIR execution into MLIR
Keep the internal runtime and JIT beside the MLIR QIR implementation. Restore their build-tree target aliases and remove cross-tree include paths.
