+++ UPGRADING.md
@@ -167,24 +167,6 @@ Boost.Multiprecision or GMP.
 
 ### QIR execution
 
-MQT Core no longer builds or installs the `mqt-core-qir-runner` executable. Use
-the DDSIM QDMI device to execute supported QIR programs with MQT Core, or use an
-external runtime such as QIR-Runner.
-
-The DDSIM QDMI device accepts a positive `int` random-number generator seed as
-custom job parameter 1 (`QDMI_DEVICE_JOB_PARAMETER_CUSTOM1`, or `custom1` in the
-Python API). This replaces the runner's `--seed` option and applies to both
-OpenQASM and QIR sampling jobs. The DDSIM device requires each QIR module to
-contain exactly one function with the standard `entry_point` attribute. Use an
-external runtime such as QIR-Runner for modules that require named entry-point
-selection.
-
-The QIR runtime and just-in-time compiler are now internal implementation
-details of the DDSIM QDMI device. MQT Core no longer provides the build-tree
-`MQT::CoreQIRRuntime` and `MQT::CoreQIRJIT` CMake aliases or the headers below
-`mqt-core/qir/runtime` and `mqt-core/qir/jit`. MQT Core does not provide a
-replacement standalone C++ API.
-
 Dynamic QIR inputs must use the current QIR 2.1 resource-management interface.
 Legacy allocator and output overloads are no longer accepted.
 
@@ -204,8 +186,7 @@ MQT Core now builds its MLIR-based compiler infrastructure unconditionally. LLVM
 including as a CMake dependency or Python package. The `BUILD_MQT_CORE_MLIR`
 CMake option has been removed. MQT Core also builds QIR support in the DDSIM
 QDMI device unconditionally, so the `BUILD_MQT_CORE_QDMI_DDSIM_WITH_QIR` option
-has been removed. The `BUILD_MQT_CORE_QIR_RUNNER` option has been removed with
-the runner. Remove all three options from build scripts and presets.
+has been removed. Remove both options from build scripts and presets.
 
 We offer pre-built distributions for all supported platforms as part of the
 `setup-mlir` project at
