+++ mlir/lib/Dialect/QIR/CMakeLists.txt
@@ -9,3 +9,4 @@
 add_subdirectory(Utils)
 add_subdirectory(Builder)
 add_subdirectory(Transforms)
+add_subdirectory(Execution)
