+++ mlir/lib/Dialect/QIR/Execution/Runtime/CMakeLists.txt
@@ -11,9 +11,10 @@ set(TARGET_NAME ${MQT_CORE_TARGET_NAME}-qir-runtime)
 if(NOT TARGET ${TARGET_NAME})
   # Add internal QIR runtime library
   add_library(${TARGET_NAME} STATIC QIR.cpp Runtime.cpp)
+  add_library(MQT::CoreQIRRuntime ALIAS ${TARGET_NAME})
 
   target_include_directories(${TARGET_NAME}
-                             PUBLIC $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/src/qir/include>)
+                             PUBLIC $<BUILD_INTERFACE:${MQT_MLIR_SOURCE_INCLUDE_DIR}>)
   set_target_properties(${TARGET_NAME} PROPERTIES POSITION_INDEPENDENT_CODE ON)
 
   # Add link libraries
